requirements.txt
pandas
google-generativeai
ipython